+++ apscheduler/schedulers/tornado.py
@@ -23,6 +23,8 @@ class TornadoScheduler(BaseScheduler):
     """
     A scheduler that runs on a Tornado IOLoop.
 
+    The default executor can run jobs based on native coroutines (``async def``).
+
     =========== ===============================================================
     ``io_loop`` Tornado IOLoop instance to use (defaults to the global IO loop)
     =========== ===============================================================
@@ -49,6 +51,10 @@ class TornadoScheduler(BaseScheduler):
         if self._timeout:
             self._ioloop.remove_timeout(self._timeout)
             del self._timeout
+
+    def _create_default_executor(self):
+        from apscheduler.executors.tornado import TornadoExecutor
+        return TornadoExecutor()
 
     @run_in_ioloop
     def wakeup(self):